+++ packages/react-native/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @agentskit/react-native
+
+## Unreleased
+
+### Patch Changes
+
+- Add `Message.contentStyle` and `InputBar.inputStyle` pass-throughs for native text theming.